+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - learning rate scheduler value plan visualizer script in [`scripts/visualize_learning_rate.py`](scripts/visualize_learning_rate.py)
 - `save_best_metrics` config to save model checkpoint on best metrics
-- `save_last_epoch` config to save last epoch model when training
+- always save checkpoint of last epoch model
 - save `best_metrics` value in model checkpoint
 
 ### Changed